+++ philo/utils/memalloc.c
@@ -6,13 +6,13 @@
 /*   By: vvoronts <vvoronts@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/03 16:44:48 by vvoronts          #+#    #+#             */
-/*   Updated: 2025/03/05 15:12:36 by vvoronts         ###   ########.fr       */
+/*   Updated: 2025/03/13 11:16:03 by vvoronts         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-void	*memalloc(size_t size, t_ctx *ctx)
+void	*memalloc(long size, t_ctx *ctx)
 {
 	void	*mem;
