+++ philo/utils/strlen.c
@@ -1,20 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_strlen.c                                        :+:      :+:    :+:   */
+/*   strlen.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: vvoronts <vvoronts@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/03 16:41:42 by vvoronts          #+#    #+#             */
-/*   Updated: 2025/03/03 16:43:22 by vvoronts         ###   ########.fr       */
+/*   Updated: 2025/03/13 11:16:59 by vvoronts         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-size_t	ft_strlen(const char *str)
+long	ft_strlen(const char *str)
 {
-	size_t	i;
+	long	i;
 
 	i = 0;
 	while (str[i])
